Implemented State, Dashboard NavBar and Sidebar
New Folders:
(components)
This folder holds the new navbar and sidebar for the homepage. It is created in this specific way so that the application does not render these as separate pages.

State
This folder holds state specific information. For example, definition of light and dark theme functions as well as collapsing sidebar functions. If you're insterested in how light and dark themes map to each other go to the tailwind config.

Old Pages:
If you want to access any of the old pages to do ur current work, you there an extra folder you have to go thorugh. For example the old settings page is now  http://localhost:3000/dashboard/oldDash/settings

Color Definition:
True black and white no longer work, instead use a customBlack and customWhite if you need to change colors using tailwind. This is due to some things with dark mode i am working on.

Dashboard Wrapper Class:
Idea is that we never have to re-render the sidebar and navbar because we are using child components.